+++ examples/flow-react-example/src/ImportHookDefault.gen.js
@@ -10,6 +10,9 @@ type $any = any;
 // flowlint-next-line nonstrict-import:off
 import {default as makeNotChecked} from './hookExample';
 
+// flowlint-next-line nonstrict-import:off
+import {default as defaultNotChecked} from './hookExample';
+
 // In case of type error, check the type of 'make' in 'ImportHookDefault.re' and './hookExample'.
 export const makeTypeChecked: ({|
   +person: person, 
@@ -23,7 +26,22 @@ export const make: mixed = function hookExample(Arg1: $any) {
   return result
 };
 
+// In case of type error, check the type of 'default' in 'ImportHookDefault.re' and './hookExample'.
+export const defaultTypeChecked: ({|
+  +person: person, 
+  +children: React$Node, 
+  +renderMe: ImportHooks_renderMe<string>
+|}) => React$Node = defaultNotChecked;
+
+// Export '$$default' early to allow circular import from the '.bs.js' file.
+export const $$default: mixed = function hookExample(Arg1: $any) {
+  const result = defaultTypeChecked({person:{name:Arg1.person[0], age:Arg1.person[1]}, children:Arg1.children, renderMe:Arg1.renderMe});
+  return result
+};
+
 // flowlint-next-line nonstrict-import:off
 import type {renderMe as ImportHooks_renderMe} from './ImportHooks.gen';
 
 export type person = {| +name: string, +age: number |};
+
+export default $$default;