+++ examples/typescript-react-example/src/ImportHookDefault.bs.js
@@ -4,8 +4,11 @@ import * as ImportHookDefaultGen from "./ImportHookDefault.gen";
 
 var make = ImportHookDefaultGen.make;
 
+var make2 = ImportHookDefaultGen.default;
+
 export {
   make ,
+  make2 ,
   
 }
 /* make Not a pure module */